sparse_ind_utk/requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
shap>=0.44.0
torch>=2.2.0
torch-geometric>=2.5.0
cvxpy>=1.4.0
scipy>=1.12.0
statsmodels>=0.14.0
matplotlib>=3.8.0
seaborn>=0.13.0
pyyaml>=6.0.1
tqdm>=4.66.0
networkx>=3.2.0
jupyter>=1.0.0
